Make yboard a global object
